#4877 Move git and svn tests
- SCM tests were moved to a new place, so both message tests need
  to be reallocated to there.
